undoing service refactor that was unneeded
